requirements-py3.txt
pbr>=0.6,!=0.7,<1.0
six>=1.7.0
PyYAML
testtools>=0.9.32
